Sample fit & finish
