+++ iothub_client/samples/pnp/pnp_temperature_controller/CMakeLists.txt
@@ -9,7 +9,6 @@ set(pnp_temperature_controller_c_files
     pnp_temperature_controller.c
     pnp_thermostat_component.c
     pnp_deviceinfo_component.c
-    ../common/pnp_device_client_ll.c
 )
 
 IF(WIN32)
