+++ iothub_client/samples/pnp/pnp_temperature_controller/pnp_deviceinfo_component.c
@@ -5,10 +5,7 @@
 
 // PnP routines
 #include "pnp_deviceinfo_component.h"
-// OLD code - moved into API #include "pnp_protocol.h"
-
 #include "iothub_properties.h"
-
 
 // Core IoT SDK utilities
 #include "azure_c_shared_utility/xlogging.h"
@@ -43,73 +40,36 @@ static const char PnPDeviceInfo_TotalStoragePropertyValue[] = "10000";
 static const char PnPDeviceInfo_TotalMemoryPropertyName[] = "totalMemory";
 static const char PnPDeviceInfo_TotalMemoryPropertyValue[] = "200";
 
-//
-// SendReportedPropertyForDeviceInformation sends a property as part of DeviceInfo component.
-//
-/* old code for sending an individual property.
-static void SendReportedPropertyForDeviceInformation(IOTHUB_DEVICE_CLIENT_LL_HANDLE deviceClientLL, const char* componentName, const char* propertyName, const char* propertyValue)
-{
-    IOTHUB_CLIENT_RESULT iothubClientResult;
-    STRING_HANDLE jsonToSend = NULL;
-
-    if ((jsonToSend = PnP_CreateReportedProperty(componentName, propertyName, propertyValue)) == NULL)
-    {
-        LogError("Unable to build reported property response for propertyName=%s, propertyValue=%s", propertyName, propertyValue);
-    }
-    else
-    {
-        const char* jsonToSendStr = STRING_c_str(jsonToSend);
-        size_t jsonToSendStrLen = strlen(jsonToSendStr);
-
-        if ((iothubClientResult = IoTHubDeviceClient_LL_SendReportedState(deviceClientLL, (const unsigned char*)jsonToSendStr, jsonToSendStrLen, NULL, NULL)) != IOTHUB_CLIENT_OK)
-        {
-            LogError("Unable to send reported state for property=%s, error=%d", propertyName, iothubClientResult);
-        }
-        else
-        {
-            LogInfo("Sending device information property to IoTHub.  propertyName=%s, propertyValue=%s", propertyName, propertyValue);
-        }
-    }
-
-    STRING_delete(jsonToSend);
-}
-*/
-
 void PnP_DeviceInfoComponent_Report_All_Properties(const char* componentName, IOTHUB_DEVICE_CLIENT_LL_HANDLE deviceClientLL)
 {
-    /* old code, this is NOT batched
-    // NOTE: It is possible to put multiple property updates into a single JSON and IoTHubDeviceClient_LL_SendReportedState invocation.
-    // This sample does not do so for clarity, though production devices should seriously consider such property update batching to
-    // save bandwidth.  PnP_CreateReportedProperty does not currently accept arrays.
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_SoftwareVersionPropertyName, PnPDeviceInfo_SoftwareVersionPropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_ManufacturerPropertyName, PnPDeviceInfo_ManufacturerPropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_ModelPropertyName, PnPDeviceInfo_ModelPropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_OsNamePropertyName, PnPDeviceInfo_OsNamePropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_ProcessorArchitecturePropertyName, PnPDeviceInfo_ProcessorArchitecturePropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_ProcessorManufacturerPropertyName, PnPDeviceInfo_ProcessorManufacturerPropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_TotalStoragePropertyName, PnPDeviceInfo_TotalStoragePropertyValue);
-    SendReportedPropertyForDeviceInformation(deviceClientLL, componentName, PnPDeviceInfo_TotalMemoryPropertyName, PnPDeviceInfo_TotalMemoryPropertyValue);
-    */
-
+    IOTHUB_CLIENT_RESULT iothubClientResult;
     unsigned char* propertiesSerialized = NULL;
     size_t propertiesSerializedLength;
     IOTHUB_REPORTED_PROPERTY properties[8] =  {
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_SoftwareVersionPropertyName, PnPDeviceInfo_SoftwareVersionPropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_ManufacturerPropertyName, PnPDeviceInfo_ManufacturerPropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_ModelPropertyName, PnPDeviceInfo_ModelPropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_OsNamePropertyName, PnPDeviceInfo_OsNamePropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_ProcessorArchitecturePropertyName, PnPDeviceInfo_ProcessorArchitecturePropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_ProcessorManufacturerPropertyName, PnPDeviceInfo_ProcessorManufacturerPropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_TotalStoragePropertyName, PnPDeviceInfo_TotalStoragePropertyValue},
-        {IOTHUB_REPORTED_PROPERTY_VERSION, PnPDeviceInfo_TotalMemoryPropertyName, PnPDeviceInfo_TotalMemoryPropertyValue}
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_SoftwareVersionPropertyName, PnPDeviceInfo_SoftwareVersionPropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_ManufacturerPropertyName, PnPDeviceInfo_ManufacturerPropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_ModelPropertyName, PnPDeviceInfo_ModelPropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_OsNamePropertyName, PnPDeviceInfo_OsNamePropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_ProcessorArchitecturePropertyName, PnPDeviceInfo_ProcessorArchitecturePropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_ProcessorManufacturerPropertyName, PnPDeviceInfo_ProcessorManufacturerPropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_TotalStoragePropertyName, PnPDeviceInfo_TotalStoragePropertyValue},
+        {IOTHUB_REPORTED_PROPERTY_VERSION_1, PnPDeviceInfo_TotalMemoryPropertyName, PnPDeviceInfo_TotalMemoryPropertyValue}
     };
 
     const int numProperties = sizeof(properties) / sizeof(properties[0]);
-    IoTHub_Serialize_ReportedProperties(properties, numProperties, componentName, &propertiesSerialized, &propertiesSerializedLength);
 
-    // Unlike original sample, everything gets batched on a single send.
-    IoTHubDeviceClient_LL_SendPropertiesAsync(deviceClientLL, propertiesSerialized, propertiesSerializedLength, NULL, NULL);
+    // The first step of reporting properties is to serialize it into an IoT Hub friendly format.  You can do this by either
+    // implementing the PnP convention for building up the correct JSON or more simply to use IoTHub_Serialize_ReportedProperties.
+    if ((iothubClientResult = IoTHub_Serialize_ReportedProperties(properties, numProperties, componentName, &propertiesSerialized, &propertiesSerializedLength)) != IOTHUB_CLIENT_OK)
+    {
+        LogError("Unable to serialize reported state, error=%d", iothubClientResult);
+    }
+    // The output of IoTHub_Serialize_ReportedProperties is sent to IoTHubDeviceClient_LL_SendPropertiesAsync to perform network I/O.
+    else if ((iothubClientResult = IoTHubDeviceClient_LL_SendPropertiesAsync(deviceClientLL, propertiesSerialized, propertiesSerializedLength, NULL, NULL)) != IOTHUB_CLIENT_OK)
+    {
+        LogError("Unable to send reported state, error=%d", iothubClientResult);
+    }
+
     free(propertiesSerialized);
 }
-
